Move Python code out of EGMO website repository.
Python code for generating a mathematical olympiad website was
formerly maintained as part of the same repository as the EGMO website
itself.  Now that this code is largely generic with EGMO-specific
configuration in separate files, it is moved to its own repository to
be maintained there separately from the website, and to be joined
later by code for the registration system and for generating documents
from data from the registration system (currently all in the
registration system repository along with EGMO-specific
configuration).