Support Python 3 in site generation.
Although the registration system cannot yet be used with Python 3
because of the lack of Python 3 support in Roundup, no such issue
applies for the other code.  This patch duly enables Python 3 support
for site generation, producing identical output for the EGMO site
whether run with Python 2 or Python 3 (although about twice as slow
with Python 3, possibly because of the decoding and encoding where
UTF-8 is used directly internally with Python 2).