Support processing data direct from registration system for document generation.
In order to use the matholymp object model for people, countries and
events to generate documents for an event yet to be held or in
progress, it is necessary to support taking data direct from files
downloaded from the registration system.  This patch adds that
support.  The main difference is that there is no CSV file describing
the current event; any information about it comes from a configuration
file instead.  The object model also needs extending to cover some
non-public information that is used in generating documents such as
name badges, so this patch also adds that information to the object
model.  A new variety of sort key is added for generating exam papers
and labels for desks, sorting first by contestant number so that all
contestants 1 come before all contestants 2, and so on.