Further enhance handling of data from registration system + config file.
Years are added to the information that can be obtained from a config
file when processing data for the current year from the registration
system together with a config file.  Other data is converted to
strings for greater consistency with how such data is handled when it
comes from a CSV file.