Support use of common code within registration system to generate CSV files.
Starting the process of using matholymp code within the registration
system, this patch adds the required support to use this code within
the registration system to generate CSV files.

A RoundupDataSource class is added to access information from within
Roundup.  The CSV generation is refactored to handle the variants
relevant for the registration system; support is added for various
fields only used there.  A RoundupSiteGenerator class handles
generating the data sent back to the user when CSV files are
requested.