Support automatic creation of user account for country.
Previously, after creating a country in the database it was necessary
to create a user account for registering people from that country, and
to send the username and password to the relevant contact, manually.
With this patch, a contact email address can be entered when creating
a country and the user account is created automatically, with a random
password generated and sent to that address.