Allow country and role names to contain commas.
Previously, this would cause "Guide For" and "Other Roles" in CSV
files of people to be misinterpreted when those files were read in.
Now, those fields are encoded and interpreted as if their value (after
CSV decoding) is itself the row of a single-row CSV file, instead of a
simple comma-separated string, so that country and role names
containing commas are appropriately encoded.