Make coordination forms more configurable.
This patch adds mo-document-generate and coord-form-template.tex
support for greater configurability of coordination forms.  A
background design may be included with the --background option; the
logo may come from such a design rather than being directly included
by the template (so allowing the unmodified template to be used in
testing, where no logo is available) and the template may allow
configuring whether headings are on the left or the right, to fit with
different background designs.