Support special prizes.
Contestants may have an arbitrary list of "Extra Awards", for both the
registration system and the static site, intended to be used for
"Special Prize" but potentially also usable if any events have
multiple kinds of special prizes each with their own description.  A
new mo-static-upgrade script is added to add this new column to
existing CSV files on the static site (and, more generally, to do any
updates to those CSV files required as part of a matholymp upgrade).
Existing tests are modified and tests of the new functionality added.