Stop mo-static-import taking event number on command line.
It is desirable for the process of importing data after an event to be
as simple and automated as possible.  To that end, remove the need to
specify an event number manually on the mo-static-import command line
by making it just use the number from the imported .csv files (which
was previously checked for consistency with the command line).  The
intent is to add support for downloading data from the registration
system, so a directory with input data does not need specifying
manually either, and for mo-static-import to obtain medal boundaries
from the registration system as well.