Give max as well as min values in cumulative statistics of scores.
The min value, the only value previously shown, is the number of
contestants with at least a given score if all unentered scores are
zero.  The max value is the number if all unentered scores are the
maximum score for that problem.  If these are different, both are now
shown, so making visible the ranges in which the final statistics
might fall once all scores are entered.