Add "Other" to default set of choices for gender.
It is increasingly widely accepted that when a gender needs to be
selected, choices should not be limited to the binary genders of
"Female" and "Male" as some people may not identify with either of
those genders.  This patch adds an "Other" option to the matholymp
registration system's default set of genders (it was already possible
to add such a gender from the command line if desired).