Support people with roles for more than one country at an event.
Normally, a person has roles with only a single country at an
olympiad.  If someone (typically from the host country) has both a
normal role such as Leader and a staff role such as on the organising
committee, this can be represented by listing them under the country
of their normal role and listing the staff role under their other
roles.  If, however, someone is Leader or Deputy Leader of more than
one team, typically A and B teams from the same country, this cannot
currently be represented.  The person can be registered multiple
times, but if the same URL for past participation is specified, this
will cause an error in static site generation when the data is
imported into the static site; if no URL for past participation is
specified, multiple entries will be created on import and if they are
merged manually, such an error will again occur.

This patch enables such arrangements to be handled in static site
generation, although it is still the case that an import without past
participation URLs for such a person will create multiple person
entries for them; those must then be merged manually (by adjusting the
person number for one of them to match that for the other, moving
photos and adjusting their URLs accordingly).  A PersonEvent is
defined to represent the participation of one Person for one Country
at one Event; the mapping from ids to people at an event now returns a
list of PersonEvents rather than one PersonEvent.