Update for Roundup 1.5.1.
add_ok_message is used instead of ok_message.append as per an API
change to avoid XSS issues (would only apply in the event of HTML
special characters in a country name).  This means the code will no
longer work with older Roundup versions because of the use of a new
API.  The Roundup patches are regenerated to apply to Roundup 1.5.1
with no line number differences.  Parts of the (very old)
issue2086536-back_postgresql.patch that are not needed with current
Roundup, some of which would no longer apply because of changes in
1.5.1, are removed.