Automatically check formats of uploaded files.
Although users generally upload files in JPEG and PNG formats,
occasionally an upload is done using some other format, which may not
work in browsers or when the image is used by LaTeX in name badge
generation.  Also, sometimes a file may be uploaded with a filename
extension that does not match its contents.  To avoid needing to check
for these issues manually, the registration system is extended to
require that flags be in PNG format and that photos be in JPEG or PNG
format, with the format of the file's contents being required to match
its filename extension.

Rather than just using the imghdr module directly in the auditors to
determine a file's format, it is wrapped in anticipation of additional
future types of uploads for which other formats, not known to the
imghdr module, may be appropriate.