Move some registration-system site-generation code out of roundupreg/.
Some registration system code purely generates output using the
general matholymp data interfaces, without using anything specific to
the Roundup-based registration system, but where the output in
question is of a form only used from the registration system not from
the static site.  In principle there could be multiple registration
system back ends, and to support that it is desirable to separate such
generic code from the code that is genuinely Roundup-specific.  Thus,
this patch moves some such code to a matholymp.regsitegen module, with
the intent that more code will move to the general interfaces and move
there in future (in particular, the generation of zip files and the
registration status page).