Rework upload filename handling.
URLs for photos and flags in CSV files now use a canonical form of the
filename rather than that used on upload (Roundup doesn't care).  ZIP
files use a canonical form of the filename extension, so
mo-static-import no longer needs to convert.  Logic relating to
filename extensions and their canonicalisation is now in fileutil.py.
Lower-casing of extensions is now done properly for ASCII rather than
potentially depending on the locale if an extension containing 'i'
were allowed.