Arrange ZIP files by person / country id, not file id.
This is a more natural arrangement for users to deal with.  It also
facilitates local filenames being information supplied through the
data interface and moving ZIP file generation to regsitegen.py, by
enabling ZIP file generation not to care about file ids which are not
an appropriate concept for the data interface.