Clarify effects of editing email addresses online.
People might think that editing an email address in the registration
system, for a user or country, has the effect of causing the new
address to be used for all subsequent contact relating to that event.
Make clear the very limited uses of email addresses in the
registration system and that organizers must be contacted separately
to change the address used for other purposes.