Support online scoreboard caching.
Generating the online scoreboard involves many database queries to
access the required information about all participants and their
countries.  During coordination, the online scoreboard can have many
people accessing it simultaneously, causing significant system load
from all the simultaneous instances of roundup-server generating the
scoreboard.  However, the number of times it changes during
coordination is much smaller than the number of times it is accessed.
Thus, it makes sense to cache the generated text so that can just be
read from a file for most scoreboard accesses, with many fewer
database accesses being involved to serve the page.

This patch implements such caching.  Accesses as a user with the Score
permission (such as an administrative user) always force regeneration
of the cached scoreboard; other accesses use it if not marked invalid.
Any change to a person, country, event or role results in the cache
being marked invalid by touching a file that is tested for when using
the cache.  Accesses to the cache are serialized with a lock file, to
avoid several simultaneous scoreboard accesses resulting in several
simultaneous regenerations if the cache is marked invalid at that
point.  This serialization only works on Unix-like operating systems;
the caching is disabled elsewhere.