Check class and node id for actions.
Previously, registration system actions just assumed they were called
for the intended class, and did not check whether a node id was
present if any such node id was ignored.  This could result in
permission checks being run with incorrect arguments (run for the
class passed when the action was then applied to the expected class),
although this was not in fact a security issue as none of the
permissions in question actually depended on class or node id.

Now, checks are systematically made for every action that it is called
for the right class, and, except where an action is OK both with and
without a node id (scores RSS feed, which can be global or for just
one country), that one is present or not as expected.