Support different country names after "in".
To use e.g. "in the United Kingdom" or "in the Netherlands" instead of
just "in United Kingdom" or "in Netherlands", while still keeping the
name without "the" in contexts such as tables where that is
appropriate.  This is only relevant to static site generation, not to
online registration.  mo-static-upgrade adds the "Country Name In"
column to the CSV file of events, and it may be edited after the
upgrade in cases where the new feature is relevant.