Use drop-down menus for day, month and year of birth.
Date of birth now uses separate String database fields for day, month
and year, instead of a Date field.  This avoids problems with people
trying to enter dates in the wrong format, or using the pop-up
calendar selector requiring clicking back year by year for many years
to select a date of birth.  The data model now uses datetime.date
objects for dates of birth.  The representation in CSV files is
unchanged.