Allow description of day as of which ages are shown to vary between events.
An Age Day Description column may be used in the CSV file of events.
Generated pages for people will have the same wording as before if the
same description applies for all events where that person was a
contestant, or a description of which descriptions apply for which
years if it varies between years when that person competed.