Support CSV file of medal boundaries from registration system.
A medal_boundaries_csv action is added to allow the exact intended
medal boundaries to be downloaded from the registration system, rather
than just inferred from the awards in people.csv.  This is
deliberately not documented, and there are no links to
event?@action=medal_boundaries_csv added to use it, since the intended
purpose is with future support for automatic downloads from
mo-static-import rather than for manual download.