Do not link ZIP files from registration system by default.
The ZIP files of flags and photos are available in the registration
system for their uses in document generation and importing data into
the static site after an event.  Such files are not provided in the
static site, where someone wishing to download flags or photos in bulk
needs to follow the links provided in the CSV files of default of
countries and people.  Since these files in the registration system
are an artifact of administrative uses of that data, and since
downloadable information from the registration system should generally
be as similar as possible to that downloadable from the static site,
this change arranges for the links to be visible only when logged in
administratively (although the ZIP files themselves, containing only
public information, remain downloadable unconditionally by URL).