Support automatic data download in mo-static-import.
As long as the static site is set up with event_active_number pointing
to the event for which the registration system is running, data from
that event can now be downloaded automatically by mo-static-import.
Medal boundaries are not yet downloaded automatically, although the
registration system does provide an action to download them.