Correct static site age descriptions when different from staticsite.cfg.
A bug caused the static site to have incorrect descriptions of the
dates as of which ages are given for contestants who competed in years
all of which had the same such description but for which that
description was not the one in staticsite.cfg.  This case was not
covered by existing testcases, so a test for it is added by this
patch.