Refactor registration system handling of uploaded files.
Computing the URL (for CSV files) of an uploaded file is refactored
into a new function db_file_url.  The class name is passed explicitly
to that function and to db_file_format_contents and db_file_extension
instead of having separate functions for 'file' and 'private_file'.
Checks on the format of uploaded files are refactored into a new
function audit_file_format.

This is in preparation for planned changes to uploaded file handling:
no longer using the 'files' name and @file special handling for flags
and photos (current development Roundup does not support @file for
Link properties; this may or may not change before 1.6 is released,
but in any case it seems cleaner to name these properties 'flag' and
'photo' to reflect what they are), and using separate 'flag', 'photo'
and 'consent_form' classes instead of 'file' and 'private_file', to
reflect how different kinds of files may have different access rules
(and in particular, in preparation for supporting specifying when a
photo is uploaded whether it is for the website or only for the name
badge).