Update for Roundup 1.6.0.
Page templates from Roundup, and config.ini, are updated for Roundup
changes.  CSRF support is added for all the matholymp-specific POST
submit buttons (four in HTML templates, one generated in Python code)
so that CSRF tokens can be required.  Obsolete or believed-obsolete
Roundup patches are removed.

No changes are made to the form of permission check functions, and no
changes are made to support a future default of False for props_only.