Handle empty filenames in registration system.
Empty filenames should not occur, but were observed in testing with
Python 3 as a consequence of a Roundup Python 3 issue.  If they do
occur, this patch ensures, to improve robustness in that case, that
db_file_extension handles them properly rather than passing them
through to file_extension which expects a string.

Given that they are not known to occur with properly functioning
browsers and released versions of Roundup, this is not considered a
user-visible fix and so does not need a release note.