Support automated tests for registration system.
Automated tests of the registration system are desirable to reduce the
extent of manual testing needed for any change, especially for
globally risky changes such as Python 3 support.

This initial patch adds a framework for such tests, using
MechanicalSoup; the tests in question are automatically skipped if
Roundup and MechanicalSoup are not available.  The tests use both the
registration system code and the example page templates shipped with
matholymp, with only the minimum changes made to them by the test
process in support of their use for testing.  Only one initial test is
added with the framework, a minimal test that the various templates
can be loaded without errors; it is expected that more substantive
tests, of both functionality and bug fixes (past and future), will be
added over time, and the test framework extended as needed to add
features required by such tests.

The structure of the tests mean that tests failing some generic
verification of responses from the registration system will produce
ERROR results from generic code raising exceptions, rather than FAIL,
as the TestCase object with assertion methods is not available in the
functions making those checks.  This is not considered a significant
problem, and in some cases may be desirable if a problem loading one
page would result in subsequent problems with later parts of the test.