Clean up and test which page templates ask to login or give access error.
Previously, all User accounts could view the gender, tshirt, language
and arrival classes; since these are only actually relevant to
accounts with the Register role, those permissions are moved to that
role.  Previously, consent_form.index.html would show an empty table
of the consent forms a user was able to view, instead of giving an
error when the user doesn't have access to the consent_form class at
all (Roundup automatically handled not listing consent forms to which
the user did not have access); this is changed to have similar errors
and conditions to other templates, and the same is done with
flag.index.html and photo.index.html although currently all flags and
photos are public.

Tests of the registration system are then made to check for each page
accessed whether it asked for login / gave an access error, and
whether this is as expected for the page in question.