Adjust registration system test handling of generated mail.
Rather than duplicating logic to read the file of generated mail (once
when producing an exception for unexpected mail, once when looking for
username/password in expected mail), always store the generated mail
in an attribute of the RoundupTestSession, and access that when the
content is needed.

Also avoid using str.decode() on the username and password in the
Python 2 case (although having Python 2 unicode object there does in
fact work for the test, it's not what's intended, and is inconsistent
with the general matholymp approach of always using the native str
type for strings.