Fix possible race condition in consent form upload.
Theoretically, the code allowed the following sequence of actions for
two registering users from different countries:

1. User 1 uploads a consent form, the person auditor passes and the
consent_form item is created.

2. User 2 (malicious) submits a person edit or creation for their
country that tries to link to the newly created consent_form item from
step 1 (linking to an existing consent_form object is not possible
from the web form, but is possible with appropriate synthetic form
submissions).  Because no country is set on the consent_form object,
the check disallowing linking to a consent form from another country
does not stop this.

3. The person reactors for the two submissions run, in either order.
If that for user 2's submission runs first, they only very briefly
have access to the consent form; if that for user 2's submission runs
second, it ends up linked to their country.

This is fixed by adding a check on the user that created a
consent_form object when linking to one without a country set.