Restrict access to superseded flags and photos.
This accords with old versions of person and country data in the
journal not being public, and likewise the pages for retired person
and country items.  By adding access control logic for photos, it also
serves as further preparation for supporting badge-only photo uploads.

Links are added from flag and photo items back to the corresponding
country and person records, set in country and person reactors, as for
consent forms, and with a check, as for consent forms, that an attempt
is not being made to link to a photo for another person or just
uploaded by another user; for consistency, the link from a
consent_form item is changed to point to a person rather than a
country (and corresponding updates are made to various logic around
consent forms).