Distinguish website and badge photos in data model.
Preparation for supporting upload of badge-only photos to the
registration system.  Both public and full versions of people.csv
include Photo URL (URL of website photo), but only full version has
Badge Photo URL, so the columns of the public version remain a subset
of those of the full version.  Uses of photo_url and photo_filename
are updated to use badge photos if appropriate; for the ZIP file of
photos from the registration system, the code now determines which
version to use based on the same logic as for the CSV file of people
(but currently all photos uploaded are both website and badge photos,
so the two versions of the ZIP file remain identical until actual
support for badge-only photos is added).