Add support for collecting additional consent information on registration.
A new configuration variable matholymp_consent_ui is added to control
whether such information is collected.  Initially, the new information
collected (if this variable is set to Yes) is whether permission is
given for publication of photos taken during the event, but the same
variable is also intended to be used to control further such questions
in future.  The answer to the new question determines the value stored
in a new Event Photos Consent column in the full version of
people.csv.