Detect invalid email addresses in country auditor.
Previously, invalid email addresses would be detected only via the
user auditor, from user creation called from the country reactor.
This still had the effect of preventing country creation with a bad
email address (commit of database changes occurs only after reactors
execute), but detecting such an issue in the auditor rather than the
reactor is logically appropriate, and means the issue is also detected
when no user creation attempt occurs because such a user already
exists.

This is intended as preparation for allowing multiple contact email
addresses to be specified for a country (so that multiple contacts
receive the email giving username and password).