Generate papers for Deputies in separate file from those for Leaders.
Papers for Deputy Leaders (and Observers with Deputy and Contestants)
are generated in a separate file from those for Leaders (and Observers
with Leader).  This makes it more convenient to distribute the papers
separately to those groups if in different places at the relevant time
(rather than needing the Leader or someone else to manually split out
the papers for the Deputy).