Include official / unofficial information in summary table of countries.
This information was previously included in the per-event tables of
countries (in the static site and registration system) but not in the
overall summary table of countries.  Now, the summary table lists
whether a country was official at the last event at which it took part
(with this being blank if official and unofficial countries were
distinguished at some events but not at that one).