Use RegisterAllCountries more consistently in registration system.
Some places in the registration system code test whether the user,
known at that point to have some edit access, is associated with a
staff or normal country, in order to tell whether they have some form
of access associated with administrative users.

Logically, such places should generally use a Roundup Permission, and
doing so also facilitates adding new kinds of users (new Roundup
Roles) that may be associated with a staff country but with different
kinds of access (for example, self-registration users to allow staff
to complete most of their registration details themselves).

This change implements that for the case of access to register
participants from all countries (as opposed to only participants from
one's own country).  The RegisterAllCountries Permission already
existed, but was only used to control the registation links present in
the sidebar, to avoid an inappropriate link showing up for scoring
users; other places are now made to use that Permission as
appropriate.  Because scoring users do have Create permission for the
person class (for the reason explained in a comment in schema.py),
there is still one logically unnecessary check of
normal_country_person in page.html, which can be removed once scoring
users no longer have that permission.