Always disallow linking to photo or consent form from another person.
This check was previously applied only for users from normal
countries; however, such linking is not a particularly sensible
operation for administrative users either, and is not in any case
possible using the normal web interface.  As it was not possible with
the normal web interface, this is not considered a user-visible change
and so is not considered to need a release note.

This completes eliminating any checks of the user's country (as
opposed to permissions) in the person auditor.