Fix entering scores with scoring user, add basic test of entering scores.
A recent change broke entering scores with a scoring user, because the
person auditor runs when scores are entered and rejected a change of
scores, for a person from another country, in the absence of the
RegisterAllCountries permission, and similarly rejected entering
scores after registration was closed in the absence of the
RegisterAnyTime permission.  This is fixed by returning early from the
auditor in the case of a scoring user only changing scores, since all
the actual checks on scores are one in the score action.

A basic test is added of entering scores (only covers initial entry of
scores, not subsequent changes, medal boundaries or any errors
entering scores), to ensure the basic functionality stays working with
both administrative and scoring users.