Do not give Create permission for person class to the Score role.
Scoring users were not actually able to register participants, but
needed the Create permission for context.country.menu to show the list
of participating countries.  By moving to a matholymp-local function
based on the matholymp string_select to generate the menu to select a
country, this permission is no longer required.

In turn, this means another use of normal_country_person in the page
templates can be removed (uses of that function mostly being suspect
in the context of new kinds of users, such as self-registration users,
that may be associated with a staff country without the permissions
tha previously implied), and that special cases in the registration
system tests, allowing [hidden] to appear for person.item.html for
scoring users, can also be removed.

The new function also filters for normal countries as the only ones
relevant for entering scores.