Pass country number not userid to registration_status_country.
This is a more natural interface and also eliminates the last use of
normal_country_person.  Thus, the normal_country_person function is
removed.

The error in registration_status_country for non-normal countries
should never actually occur because users from such countries should
either be Omnivident and so see the general registration status page
for all countries, or not have the RegistrationStatus permission at
all.