Track expected numbers of participants / single rooms for each country.
New properties are added for countries in the registration system to
track the expected number of participants in each role, and the
expected number of single room requests.  These expected numbers are
included in the full version of the CSV file of countries, and the
registration page reflects these expected numbers when reporting roles
that have not yet been registered for a country.

This change implements all the handling of such expected numbers
needed for them to be tracked and edited by administrative users.  A
followup change is intended to enable a country's expected numbers to
be edited by registering users for that country rather than just by
administrative users, so implementing an explicit preregistration
phase where countries indicate how many participants and single room
requests they intend to register.