Add EditCountry permission.
This controls which parts of country.item.html get shown.  This is
preparation for giving registering users the permission to edit, for
their own country, details of the expected numbers of participants in
different roles and the expected number of single room requests (via a
separate page template); giving this limited ability to edit the
country class should not thereby result in various additional edit
interfaces and associated text appearing when the default country
template is viewed.