Allow registering users to edit their own preregistration data.
A new template is added for such users to edit just the expected
numbers of participants, without being able to edit other country data
for their own country or to see or edit that data for other countries;
registering users are given the appropriate permissions for such
editing.

With this change, the preregistration data can be edited at any time.
A followup change will allow controlling when it can be edited,
independent of when normal registration is permitted, and will change
the error messages about registration being disabled in the case where
preregistration is enabled (on the basis that that is a
preregistration phase before normal registration is enabled).