Adjust registration status message for multiple people with a role missing.
If multiple people expected to be registered with a given (observer)
role from a given country have not yet been registered, make the
registration page say how many have yet to be registered rather than
repeating the name of the role in the list.  This is more friendly if
the numbers are large, and avoids a potential denial-of-service that
might otherwise have occurred with very large numbers specified.