Support disabling preregistration independently from registration.
This allows for a period before registration starts when only
preregistration is open; for preregistration in parallel with
registration; for preregistration to be disabled while registration is
still open; and for both to be closed as the event approached.
Messages adapt appropriately for the different cases; the email
template needs to be edited accordingly if new countries are
registered both before and after registration opens.