Support single room requests through registration system.
Support is added for a "room type" class (e.g. single and shared
rooms) and associated specification of a room type for a person, with
restrictions on permitted room types for each role and a default room
type being specified for each role.  Administrative users can set a
room type for a person not otherwise permitted for their role, in case
there are special-case reasons for doing so.  Room types are shown on
the page for online editing of room numbers.

A subsequent change will add a text field for requesting to share with
a particular other person.