Configure badge backgrounds through database.
A new badge_type class is added in the registration system, with each
role having a link to a badge_type.  This replaces hardcoded logic in
docgen.py for mapping a role name to a badge background.  This allows
more flexibility for events to choose backgrounds for each role, and
is also intended as preparation for configuring badge colours for each
badge type through the database as well, replacing having separate
background PDFs for each badge type (so different badge types could
then have different PDFs, but would generally have a common PDF and
different colours).