Handle participant names as in passports.
The intended use of this information is for visa invitation letters
(with an intended future feature to generate such letters from the
registration system).  Thus, it is only collected when both
nationality and passport number are also collected.  This information
is optional and names as in passports are assumed to be the same as
names for the website and name badges if not otherwise specified.