Ensure only a single event object is present in registration system.
There is only meant to be a single instance of the event class in the
registration system, storing data such as medal boundaries and whether
registration is enabled; code using that class hardcodes the id '1'
for the instance being used.  This change disallows creating another
instance, and removes the 'year' property from the class as not
actually used for anything and not needed when different instances
don't need to be distinguished.  As an internal change rather than
anything user-visible (unless using a not-linked-to URL as admin to
try to create another event object), there is no release note
specifically for this change.