Support bulk registration of countries.
Support is added for registering countries in bulk by uploading a CSV
file; such a CSV file of countries and contact details can be
maintained from year to year and thus substantially reduce the manual
work involved in setting up the registration system for each year's
event.  The infrastructure for bulk registration is designed so that
much of it can also be shared by bulk registration of people for the
staff country.