Properly encode non-ASCII registration system emails.
The character set needs to be specified explicitly when setting a
message payload for the resulting message body to be properly encoded
rather than sent as 8-bit content.  Sending as 8-bit content appears
to work in most cases, but fails when Roundup is running under an
ASCII locale.

Proper encoding in turn requires the tests to handle base64-decoding
of the resulting emails.

The failures are a consequence of use of Python 3, thus not thought to
be present in a release and not needing a release note.