Fix templating errors calling html.escape on None room_share_with.
room_share_with can be None so that needs to be handled in producing
the room-edit page.  Bulk/self-registration of people will also result
in gender and room_type being able to be None there as well, so the
html.escape calls are adjusted for those.

No release notes as these are only issues with new changes since the
last release; test coverage to be added when tests are added more
generally for the room-editing page.