Properly support countries with codes or names substrings of those for other countries.
An incorrect use of Roundup's "filter" means that creation of such
countries (or editing when both were present) was disallowed, but they
could be created in the other order.  Because the auditor check on
country names is new since the last release, the release note only
mentions this as an issue for country codes.