Support self-registration accounts.
These are accounts tied to a specific registered person, that can edit
most data for that person that normal registering accounts can edit,
but not country or role; as they are intended for staff, they can also
edit phone numbers.  Such accounts cannot create person records and
have no spoecial access to those for other people.  Editing will only
work if the account's country is the same as the person's country.

In this implementation, there is no friendly user interface to create
such an account; a person number needs to be entered manually on the
user edit form.  The intended way to create such accounts is through
bulk registration of people, once that is implemented.