Use tempfile.TemporaryDirectory for automatic cleanup.
Replaces uses of tempfile.mkdtemp that had to be paired with an
explicit shutil.rmtree call.  There are some cleanup calls in places
to make this cleanup explicit, but those are not strictly required any
more.

A mkdtemp call is left in a case where automatic cleanup is not needed
(where the directory is a subdirectory of another one that does have
automatic cleanup).