Support online name badge generation.
When the registration system is configured with
matholymp_docgen_directory pointing to the document generation
directory, a "Generate name badge" button appears on each
participant's page, to allow name badges to be generated online rather
than only from the command line.  (Bulk generation of all name badges
online is to be added separately.)  LaTeX errors from name badge
generation are both sent back to the web browser and emailed to the
admin.